+++ net/minecraft/world/level/newbiome/context/LazyAreaContext.java
@@ -55,7 +55,7 @@ implements BigContext<LazyArea> {
 
     @Override
     public int nextRandom(int i) {
-        int j = (int)Math.floorMod(this.rval >> 24, (long)i);
+        int j = Math.floorMod(this.rval >> 24, i);
         this.rval = LinearCongruentialGenerator.next(this.rval, this.seed);
         return j;
     }
